test(daemon): co-locate ComposeSemanticsExtension test with the moved class
The retargeting to {Android, Desktop} left the old daemon/android
ComposeSemanticsExtensionTest asserting targets == {Android}, failing the
Android suite. Delete it and fold its coverage (after-capture hook, Capture
phase, and the process fail-fast on a missing semantics root) into the
connector-module test that lives next to the moved class, using the shared
RenderArtifactContextKeys.
